Migrate routes to new caldir-core
