+++ src/rpc/bindings.ts
@@ -6,7 +6,7 @@ type TAURI_CHANNEL<T> = (response: T) => void
 
 export type Calendar = { slug: string; name: string | null; color: string | null; provider: string | null; account: string | null; read_only: boolean | null }
 
-export type CalendarEvent = { id: string; recurring_event_id: string | null; summary: string; description: string | null; location: string | null; start: RpcEventTime; end: RpcEventTime; status: string; recurrence: Recurrence | null; master_recurrence: Recurrence | null; reminders: number[]; organizer: EventAttendee | null; attendees: EventAttendee[]; conference_url: string | null; calendar_slug: string; color: string | null; 
+export type CalendarEvent = { id: string; recurring_event_id: string | null; summary: string; description: string | null; location: string | null; start: RpcEventTime; end: RpcEventTime; status: string; recurrence: RpcRecurrence | null; master_recurrence: RpcRecurrence | null; reminders: number[]; organizer: EventAttendee | null; attendees: EventAttendee[]; conference_url: string | null; calendar_slug: string; color: string | null; 
 /**
  * RFC 3339 timestamp of the event's last modification (DTSTAMP/LAST-MODIFIED).
  * Used by the frontend to cheaply detect content changes for reload dedup.
@@ -16,7 +16,7 @@ updated: string | null }
 /**
  * Input for creating an event
  */
-export type CreateEventInput = { calendar_slug: string; summary: string; description: string | null; location: string | null; start: RpcEventTime; end: RpcEventTime; recurrence: Recurrence | null; reminders: number[] }
+export type CreateEventInput = { calendar_slug: string; summary: string; description: string | null; location: string | null; start: RpcEventTime; end: RpcEventTime; recurrence: RpcRecurrence | null; reminders: number[] }
 
 export type CredentialFieldInput = { id: string; value: string }
 
@@ -31,8 +31,6 @@ export type ProviderConnectStepKind = "oauth_redirect" | "hosted_oauth" | "crede
 export type ProviderField = { id: string; label: string; field_type: ProviderFieldType; required: boolean; help: string | null }
 
 export type ProviderFieldType = "text" | "password" | "url"
-
-export type Recurrence = { rrule: string; exdates: RpcEventTime[] }
 
 export type ResponseStatus = "accepted" | "declined" | "tentative" | "needs-action"
 
@@ -59,6 +57,8 @@ export type RpcEventTime =
  */
 { kind: "datetime_zoned"; wallclock: string; tzid: string }
 
+export type RpcRecurrence = { rrule: string; exdates: RpcEventTime[] }
+
 /**
  * Input for splitting a recurring series at a given instance.
  * 
@@ -83,7 +83,7 @@ split_end: RpcEventTime;
 /**
  * Recurrence rule for the new series. None means a single non-recurring event.
  */
-new_recurrence: Recurrence | null }
+new_recurrence: RpcRecurrence | null }
 
 export type SyncPreview = { calendar_slug: string; to_push_count: number; to_push_delete_count: number; to_pull_count: number }
 
@@ -96,7 +96,7 @@ export type UpdateEventInput = { id: string; calendar_slug: string;
 /**
  * If set and different from calendar_slug, moves the event to this calendar
  */
-new_calendar_slug: string | null; summary: string; description: string | null; location: string | null; start: RpcEventTime; end: RpcEventTime; recurrence: Recurrence | null; reminders: number[] }
+new_calendar_slug: string | null; summary: string; description: string | null; location: string | null; start: RpcEventTime; end: RpcEventTime; recurrence: RpcRecurrence | null; reminders: number[] }
 
 const ARGS_MAP = { 'caldir':'{"connect_provider":["provider_name"],"connect_provider_with_credentials":["provider_name","credentials"],"create_event":["input"],"create_local_calendar":["name","color"],"delete_event":["calendar_slug","event_id"],"delete_recurring_series":["calendar_slug","uid"],"discard":["calendar_slugs"],"get_calendar_dir":[],"get_default_calendar":[],"get_default_reminders":[],"get_event":["calendar_slug","event_id"],"get_provider_connect_info":["provider_name"],"get_time_format":[],"list_calendars":[],"list_events":["calendar_slugs","start","end"],"list_invites":["calendar_slugs"],"list_providers":[],"rsvp":["calendar_slug","event_id","response"],"search_events":["calendar_slugs","query"],"set_calendar_dir":["path"],"set_default_calendar":["slug"],"set_default_reminders":["minutes"],"set_time_format":["time_format"],"split_recurring_series_at":["input"],"sync":["calendar_slugs","allow_mass_delete"],"sync_preview":["calendar_slugs"],"update_event":["input"]}', 'config':'{"get_auto_sync_enabled":[],"get_notifications_enabled":[],"get_theme":[],"set_auto_sync_enabled":["enabled"],"set_notifications_enabled":["enabled"],"set_theme":["theme"]}', 'omarchy':'{"get_colors":[]}', 'platform':'{"needs_native_decorations":[]}' }
 export type Router = { "caldir": {connect_provider: (providerName: string) => Promise<Calendar[]>,
